+++ tfmodels/utilities/ops.py
@@ -103,11 +103,11 @@ def deconv(features, n_kernel, upsample_rate=2, k_size=4, pad='SAME', var_scope=
         return out
 
 ## BUG batch norm with training
-def batch_norm(features, reuse=False, training=True, var_scope='batch_norm'):
+def batch_norm(features, momentum=0.99, reuse=False, training=True, var_scope='batch_norm'):
     with tf.variable_scope(var_scope) as scope:
         ## ReLU
         out = tf.layers.batch_normalization(features, renorm=True, reuse=reuse,
-            momentum=0.99,
+            momentum=momentum,
             fused=None,
             training=training)
         return out
